+++ clients/admin-ui/src/features/manual-tasks/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useManualTaskColumns } from "./useManualTaskColumns";
